Final changes

+++ NormalizingFlows/src/data/density/processed_data.py
@@ -1,4 +1,6 @@
 from .data import Dataset
+import h5py
+import torch
 
 class ProcessedData(Dataset):
     def __init__(self):
@@ -7,9 +9,9 @@ class ProcessedData(Dataset):
     def _get_preprocessed(self, datapath):
         datafile = h5py.File(datapath, 'r')
 
-        self.train_data = torch.tensor(datafile['train'][:])
-        self.validation_data = torch.tensor(datafile['validation'][:])
-        self.test_data = torch.tensor(datafile['test'][:])
+        self.train_data = torch.tensor(datafile['train'][:], dtype=torch.float)
+        self.validation_data = torch.tensor(datafile['validation'][:], dtype=torch.float)
+        self.test_data = torch.tensor(datafile['test'][:], dtype=torch.float)
 
         self.train_n, self.dim_input = self.train_data.size()
         self.valid_n = self.validation_data.size()[0]
